Add custom value comparators to dictmatch
Adds a new value_cmp_func kwarg to DictNamespace.match(),
which allows users to set a custom function for comparing
all values in a dict (versus adding a custom function for
comparing a value for a specific key). Built-in functions are defined
in comparison.CompareFunctions enum.

Adds UTs for dict.match covering the new functionality.

Also update FixNamespace.match for typed vs untyped FIX matches.
With untyped or mixed, values are compared as strings. Else for
typed FIX messages the types will be checked explicitly.