Replace report_all flag with report_mode option
Instead of using a report_all flag to control reporting of
dict and fix match results, add a report_mode option.
This can be set to one of the values of the new ReportOptions
enum to control the reporting behaviour. Current reporting
options are ALL (default), NO_IGNORED or FAILS_ONLY but more
could be added.