Fix serialitation edge cases for Task results
Fix serialisation of result types so that only the "safe" builtin types
such as int, str, NoneType etc. are directly pickled, all other objects
are converted to their string representation. In particular, types that
inherit from a "safe" builtin are no longer directly pickled - this was
causing problems with Boost.Python.enum type.

Also add a simple check to prevent scheduling of tasks from the
__main__ module for Process pool - this will always fail since __main__
is a completely different module for the child worker processes.