Fix serialization edge case and rename test/ -> tests/
Fixes an edge case with serialization for result types
that inherit from a "safe" builtin such as int but may
not be safe to serialize. An example of such a type
the Boost.python.enum enumeration type.

Also renames test/ dir to tests/ to avoid name clashing
with the python builtin "test".