Unit test message flow between Pool and workers
Add a unit test to test the Pool base class in isolation from its
workers. The test sends in requests from a mock worker it controls and
asserts the responses are correct. For now this only checks a very basic
mainline flow where a worker pulls a task from the Pool and sends back
the results, but it can be extended to allow testing of more interesting
window conditions.

Fix a bug in the default queue communication so that the Pool does not
block when getting the next request if there is none.

Also adds an optional timeout to the interruptible_join thread utility,
so that we don't block forever when stopping if the Pool's main work
loop blocks for any reason.