Improve LogMatcher performance for large logfiles
Do not sleep when iterating through lines in a logfile - only sleep
after the last line is read and we are waiting for more to be produced.
Modifies the unit test to use real (temporary) logfiles instead of
mocking the open() builtin and add a testcase for a realistically sized
logfile.