Implement prototype interactiv navigation
Add components for navigating the report in interactive mode. Each report
entry has a "play" button that can be used to run the test. Since this is
a prototype, we don't yet make any API calls to the backend to run the tests,
instead a fake interactive report is used. For demonstration purposes only,
the tests are assumed to pass after 3 seconds have passed.