Add parent_uids to report entries
Adds new parent_uids attribute to all report entries.
This is a list of the UIDs of all parent entries in the report
tree from root (first element) downwards. Knowing both the UID
of an entry and the UIDs of its parents allows an entry to be
found in the report tree in O(log(n)) time.

parent_uids are automatically updated whenever a child report entry
is appended to its parent.

This new attribute is used to find and update report entries in the
interactive UI, instead of relying on the current selection state.