Use partial updates of interactive report
Enhance the backend to provide a hash of each node in the report tree.
This hash value is used to determine if there have been changes to that
particular node or in the sub-tree below it.

Compare the hash value in the frontend with previous values when
performing updates, and only drill down into parts of the report tree
where the hash value has changed. This greatly reduces the number of
HTTP requests made when polling for report updates.