Log async errors and fix report initialisation
Use the concurrent.futures package instead of multiprocessing and add a
callback to log exceptions that occur asynchronously. Add python2
backport as a dependency.

Fix and simplify report initialisation to use the dry_run method of
tests. This ensures that the report skeleton for parametrized testcases
is generated correctly and reduces code duplication.

Further fixes for running parametrized testcases in interactive mode are
required and will be made soon...