Refactor MultiTest into reusable component methods
Refactor MultiTest methods into smaller, more composable methods. The
driving goal of this change is to improve the interaction between
MultiTest and the interactive runner to remove bugs in the interactive
mode, but I hope it also provides a readability improvement and possibly
remove some bugs in MultiTest.

Replace the home-grown thread pool implementation that was jammed into
MultiTest previously and just use concurrent.futures package for
multithreading testcases that belong to an execution group.