Add a button to reset testcase state in interactive UI
Adds a new button into the toolbar for interactive mode, which triggers
report state to be reset. Since most of the state of MultiTests and suites
derives from that of testcases, and testcases derive their state from
their entries, it is enough to simply update the testcase entries to
contain no assertion entries.

Includes a refactor of the Toolbar component, to split the massive 150+
line render method into sensible sub-functions, and to add a new prop to
allow new buttons to be injected into the toolbar just for interactive
mode (so that this reset button doesn't appear for batch reports).