Set parent_uids correctly in interactive report
When merging testcase entries into the report tree, ensure that the
parent_uids attribute is set correctly. There have been some strange UI
bugs because of this not being correct.

Also, when validating a PUT request, check that the UID of the updated
entry definitely matches with the current entry - changing a UID is not
allowed.