Quick fix for test_app for windows
if exception is thrown from start, the app object still
holds the stderr and stdout files. Depends on timing of gc
runtest fixture cleanup may try to delete the file before it
is closed, which result an error on windows. Make sure we
stop the app which will close the files in the testcase.