use abspath to enhance the sys.path in tmp import
from python 3.8 cpython has a caching in import tools that is atacching
a cash for all member on the sys.path (and keep that even if an entry
is removed) Even if the path is relative the cache inside caching based
on the cwd at the time cache is created.

if one cwd to a temp dir, and add something relative to the sys.path and
load a module from there then the temp dir get into a the cache

now if the temp dir is deleted and the same thing is done again with the
same relative path, the cache would try to work on the previous tempdir
which is now resolved.

This is a real edgecase I was seeing this in a testcase where the
usecase is probbaly far from a normal application. Using absolute path
instead will make the cache refering to the whole temp path, and new
cache is created for the two run.